+++ server/controllers/history.js
@@ -38,7 +38,8 @@ module.exports = {
   },
 
   wipeHistory: async (req, res, next) => {
-    await History.collection.drop();
+    const count = await History.count({});
+    if (count) { await History.collection.drop(); }
     res.status(200).json({ success: true });
     // console.log('Deleted History');
   }
